pyproject.toml
[project]
name = "mlwiki"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.0",
    "pydantic>=2.9.0",
    "uvicorn>=0.30.6",
    "coloredlogs>=15.0.1",
    "sqlalchemy>=2.0.34",
    "cmarkgfm>=2024.1.14",
    "openai>=1.44.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "fire>=0.6.0",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
